Implement package building hooks.
A new configuration file (ql-to-deb.ini) allows specifying hooks. Those
hooks can be used to customize some stages of the package preparation
and building.

Currently only one such hook is implemented, the "orig" hook, that is
run with the Quicklisp archive file as input and is meant to deliver the
debian orig tarball as an output.